scraping/requirements.txt
scrapy
beautifulsoup4
requests
pyspark
opencv-python-headless
librosa
transformers
kafka-python
boto3
free-proxy
fake-useragent
scrapy-rotating-proxies